Update template.php
When i tested the dgd7 module on local dev drupal site, error message was produced. It stated that a variable in html.tpl.php was undefined. I visited the website linclark.com referenced in the template.php file and found you missed two lines from her code for the 'rdf_version' variable. The error seems to have disappeared. It is a good theme. I hope I have polished it slightly.